+++ Database/ClassifiedsDatabase/AppSetting-KandoBay.sql
@@ -27,8 +27,8 @@ execute temp_createAppSetting @Key = 'SearchMaxPagedRequests', @Setting = '100'
 execute temp_createAppSetting @Key = 'MaxOnlineImages', @Setting = '5'
 execute temp_createAppSetting @Key = 'SupportNotificationAccounts', @Setting = 'support@paramountit.com.au'
 execute temp_createAppSetting @Key = 'AdDurationDays', @Setting = '730', @Force = 1
-execute temp_createAppSetting @Key = 'ClientAddress', @Setting = 'PO Box 333, Lennox St,Richmond,VIC,3121,Australia', @Force = 1
-execute temp_createAppSetting @Key = 'ClientAddressLatLong', @Setting = '-37.818635,145.001470', @Force = 1
+execute temp_createAppSetting @Key = 'ClientAddress', @Setting = '47 Coven,Avenue,Heathmont,VIC,3135,Australia', @Force = 1
+execute temp_createAppSetting @Key = 'ClientAddressLatLong', @Setting = '-37.827868,145.258419', @Force = 1
 execute temp_createAppSetting @Key = 'EnableTwoFactorAuth', @Setting = 'true', @Force = 1
 execute temp_createAppSetting @Key = 'ClientName', @Setting = 'KandoBay', @Force = 1
 execute temp_createAppSetting @Key = 'EventTicketReservationExpiryMinutes', @Setting = '20'