+++ Paramount.Betterclassifieds.Presentation/ViewModels/AddressViewModel.cs
@@ -2,12 +2,11 @@
 {
     public class AddressViewModel
     {
-        public string AddressLine1 { get; set; }
-        public string AddressLine2 { get; set; }
+        public string StreetNumber { get; set; }
+        public string StreetName { get; set; }
         public string Suburb { get; set; }
         public string State { get; set; }
         public string Postcode { get; set; }
         public string Country { get; set; }
-        public string PhoneNumber { get; set; }
     }
 }