+++ Paramount.Betterclassifieds.Business/Events/TicketPrice.cs
@@ -2,8 +2,20 @@ namespace Paramount.Betterclassifieds.Business.Events
 {
     public class TicketPrice
     {
-        public decimal OriginalPrice { get; set; }
-        public decimal PriceIncludingFee { get; set; }
-        public decimal Fee { get; set; }
+        public TicketPrice()
+        {
+            // All the decimals should obtain the zero default values
+        }
+
+        public TicketPrice(decimal originalPrice, decimal priceIncludingFee, decimal fee)
+        {
+            OriginalPrice = originalPrice;
+            PriceIncludingFee = priceIncludingFee;
+            Fee = fee;
+        }
+
+        public decimal OriginalPrice { get; private set; }
+        public decimal PriceIncludingFee { get; private set; }
+        public decimal Fee { get; private set; }
     }
 }